+++ packages/react-charting/src/components/GaugeChart/GaugeChart.base.tsx
@@ -10,9 +10,9 @@ import {
   GaugeChartVariant,
 } from './GaugeChart.types';
 import { IProcessedStyleSet } from '@fluentui/react/lib/Styling';
+import { convertToLocaleString } from '../../utilities/locale-util';
 import {
   Points,
-  convertToLocaleString,
   formatValueWithSIPrefix,
   getAccessibleDataObject,
   getColorFromToken,