+++ packages/react-components/react-aria/src/activedescendant/useActiveDescendant.ts
@@ -13,6 +13,19 @@ export function useActiveDescendant<TActiveParentElement extends HTMLElement, TL
   const focusVisibleRef = React.useRef(false);
   const activeIdRef = React.useRef<string | null>(null);
   const activeParentRef = React.useRef<TActiveParentElement>(null);
+  const attributeVisibilityRef = React.useRef(true);
+
+  const removeAttribute = React.useCallback(() => {
+    activeParentRef.current?.removeAttribute('aria-activedescendant');
+  }, []);
+  const setAttribute = React.useCallback((id?: string) => {
+    if (id) {
+      activeIdRef.current = id;
+    }
+    if (attributeVisibilityRef.current && activeIdRef.current) {
+      activeParentRef.current?.setAttribute('aria-activedescendant', activeIdRef.current);
+    }
+  }, []);
 
   useOnKeyboardNavigationChange(isNavigatingWithKeyboard => {
     focusVisibleRef.current = isNavigatingWithKeyboard;
@@ -42,9 +55,9 @@ export function useActiveDescendant<TActiveParentElement extends HTMLElement, TL
       active.removeAttribute(ACTIVEDESCENDANT_FOCUSVISIBLE_ATTRIBUTE);
     }
 
-    activeParentRef.current?.removeAttribute('aria-activedescendant');
+    removeAttribute();
     activeIdRef.current = null;
-  }, [activeParentRef, getActiveDescendant]);
+  }, [getActiveDescendant, removeAttribute]);
 
   const focusActiveDescendant = React.useCallback(
     (nextActive: HTMLElement | null) => {
@@ -55,15 +68,14 @@ export function useActiveDescendant<TActiveParentElement extends HTMLElement, TL
       blurActiveDescendant();
 
       scrollIntoView(nextActive, listboxRef.current);
-      activeParentRef.current?.setAttribute('aria-activedescendant', nextActive.id);
-      activeIdRef.current = nextActive.id;
+      setAttribute(nextActive.id);
       nextActive.setAttribute(ACTIVEDESCENDANT_ATTRIBUTE, '');
 
       if (focusVisibleRef.current) {
         nextActive.setAttribute(ACTIVEDESCENDANT_FOCUSVISIBLE_ATTRIBUTE, '');
       }
     },
-    [activeParentRef, listboxRef, blurActiveDescendant],
+    [listboxRef, blurActiveDescendant, setAttribute],
   );
 
   const controller: ActiveDescendantImperativeRef = React.useMemo(
@@ -139,8 +151,24 @@ export function useActiveDescendant<TActiveParentElement extends HTMLElement, TL
 
         return target?.id;
       },
+      showAttributes() {
+        attributeVisibilityRef.current = true;
+        setAttribute();
+      },
+      hideAttributes() {
+        attributeVisibilityRef.current = false;
+        removeAttribute();
+      },
     }),
-    [optionWalker, listboxRef, focusActiveDescendant, blurActiveDescendant, getActiveDescendant],
+    [
+      optionWalker,
+      listboxRef,
+      setAttribute,
+      removeAttribute,
+      focusActiveDescendant,
+      blurActiveDescendant,
+      getActiveDescendant,
+    ],
   );
 
   React.useImperativeHandle(imperativeRef, () => controller);