+++ packages/react-components/react-rating-preview/stories/Rating/RatingShape.stories.tsx
@@ -15,8 +15,8 @@ export const Shape = () => {
   const styles = useStyles();
   return (
     <div className={styles.root}>
-      <Rating iconFilled={<CircleFilled />} iconOutline={<CircleRegular />} step={0.5} />
-      <Rating iconFilled={<SquareFilled />} iconOutline={<SquareRegular />} step={0.5} />
+      <Rating iconFilled={CircleFilled} iconOutline={CircleRegular} step={0.5} />
+      <Rating iconFilled={SquareFilled} iconOutline={SquareRegular} step={0.5} />
     </div>
   );
 };