+++ packages/react-monaco-editor/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-monaco-editor
 
-This log was last generated on Tue, 27 Feb 2024 07:25:42 GMT and should not be manually modified.
+This log was last generated on Mon, 04 Mar 2024 07:26:19 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.7.181](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.7.181)
+
+Mon, 04 Mar 2024 07:26:19 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.7.180..@fluentui/react-monaco-editor_v1.7.181)
+
+### Patches
+
+- Bump @fluentui/react-charting to v5.19.31 ([PR #30608](https://github.com/microsoft/fluentui/pull/30608) by beachball)
 
 ## [1.7.180](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.7.180)